adjusted login page
